+++ src/Attributes/Cast.php
@@ -7,7 +7,7 @@ namespace StdOut\SimpleDataObjects\Attributes;
 use Attribute;
 use StdOut\SimpleDataObjects\Contracts\CastsValue;
 
-#[Attribute(Attribute::TARGET_PARAMETER)]
+#[Attribute(Attribute::TARGET_PARAMETER | Attribute::TARGET_PROPERTY)]
 final class Cast
 {
     public function __construct(