+++ src/Attributes/IgnoreIfNull.php
@@ -6,5 +6,5 @@ namespace StdOut\SimpleDataObjects\Attributes;
 
 use Attribute;
 
-#[Attribute(Attribute::TARGET_PARAMETER)]
+#[Attribute(Attribute::TARGET_PARAMETER | Attribute::TARGET_PROPERTY)]
 final class IgnoreIfNull {}